Fix nachocove/qa#274.
Remove the swizzling code of UIAlertView from the library. (But keeping the source file just in case.) The reason
is that: 1) it is not needed anymore. (Its functionality is replaced by NcAlertView), 2) keeping it causes problem.
Specifically, dealloc is overridden and it still calls ncSetDelegate which causes an infinite recursion.